#414#Fix bugs when user navigate between pages
